Update README.md
This template provides a simple user interface with an address bar and a web view area. When the user enters a URL in the address bar and clicks the "Go" button, the loadPage() JavaScript function will attempt to navigate to the specified URL.

Remember, this is just a basic example of a browser-like interface using HTML and CSS. To create a fully functional browser, you will need to implement various features like tab management, history, bookmarks, rendering engine integration, and handle security considerations carefully. Developing a real browser is a complex task that involves multiple technologies and components.